backend/apps/data_ingestion/requirements.txt
# Data Ingestion Service - Phase 6 Dependencies

# IBKR connectivity
ib-insync==0.9.86

# Async support
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Data stores
redis[hiredis]==5.0.1
clickhouse-driver==0.2.6

# NATS messaging
nats-py==2.6.0

# HTTP server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Data manipulation
pandas==2.1.3
numpy==1.26.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging
structlog==23.2.0

# Retry logic
tenacity==8.2.3

# HTTP client for FRED/Crypto APIs
httpx==0.25.2

# JSON parsing
orjson==3.9.10
